# resources/swagger.yaml
swagger: '2.0'
info:
  version: "1.0.0"
  title: "Delilah Restó"
  description: "This is a server for handling orders to a restaurant. It has a login and billing system, its use is completely free and it has the ability to add new features if necessary."
  contact:
    name: "API Support"
    url: "https://github.com/poved4/Resto"
host: "localhost:3000"
basePath: "/resto.com"
...
schemes:
- "http"
paths:
...
  /productos/{productId}:
    get:
      tags:
      - "productos"
      summary: "Find product by ID"
      description: "This can only be done by the logged in user."
      operationId: "getProductById"
      produces:
      - "application/json"
      parameters:
      - name: "productId"
        in: "path"
        description: "The object id that need to be consulted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/products"
        "404":
          description: "Not Found"
    patch:
      tags:
      - "productos"
      summary: "Updated product"
      description: "This can only be done by manager user."
      operationId: "updateProduct"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "productId"
        in: "path"
        description: "product id to updated"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "data to update" 
        required: true
        schema:
          type: "object"
          properties:
              price:
                type: "integer"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "Invalid data supplied"
        "401":
          description: "Unauthorized"
    delete:
      tags:
      - "productos"
      summary: "Delete Product"
      description: "This can only be done by manager user."
      operationId: "deleteProduct"
      produces:
      - "application/json"
      parameters:
      - name: "productId"
        in: "path"
        description: "product id to deleted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "204":
          description: "No Content"
        "400":
          description: "Invalid data supplied"
        "401":
          description: "Unauthorized"
  /pedidos:
    get:
      tags:
      - "pedidos"
      summary: "Returns orders on inventories"
      description: "This can only be done by manager user."
      operationId: "GetAllOrders"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/orderCreated"
        "404":
          description: "Not Found"
        "401":
          description: "Unauthorized"
    post:
      tags:
      - "pedidos"
      summary: "create a new order"
      description: "This can't be done by the logged in user."
      operationId: "OrderUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "order object"
        required: true
        schema:
          $ref: '#/definitions/order'
      responses:
        '200':
          description: "successful operation"
        '400':
         description: "Invalid data supplied"
    patch:
      tags:
      - "pedidos"
      summary: "Update order state "
      description: "This can only be done by manager user."
      operationId: "updateOrderState"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "data to update" 
        required: true
        schema:
            type: "object"
            properties:
              id: 
                type: "integer"
              state:
                type: "string"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "Invalid data supplied"
        "401":
          description: "Unauthorized"
...
definitions:
...
  products:
    type: "object"
    properties:
      products_id: 
        type: "integer"
      products_Price:
        type: "integer"
      products_Name: 
        type: "string"
      products_link_img: 
        type: "string"
...
  order:
    type: "object"
    properties:
      wayToPay: 
        type: "string"
      products:
        type: "object"
        properties:
          id: 
            type: "integer"
          quantity:
            type: "integer"
  orderCreated:
    type: "object"
    properties:
      order_id: 
        type: "integer"
      fk_user_id: 
        type: "integer"
      order_state: 
        type: "string"
      order_date: 
        type: "string"
      order_price: 
        type: "string"
      order_way_to_pay: 
        type: "string"
      order_description:
        type: "string"
      products:
        type: "object"
        $ref: "#/definitions/products"